# Conflicts:
#	src/test/scala/b4processor/modules/executor/ExecutorTest.scala
#	src/test/scala/b4processor/utils/ALUValue.scala
#	src/test/scala/b4processor/utils/ReservationValue.scala

+++ src/test/scala/b4processor/utils/ReservationValue.scala
@@ -2,9 +2,9 @@ package b4processor.utils
 
 case class ReservationValue(valid: Boolean,
                             destinationTag: Int,
-                            value1: Int,
-                            value2: Int,
+                            value1: BigInt,
+                            value2: BigInt,
                             function3: Int,
                             immediateOrFunction7: Int,
                             opcode: Int,
-                            programCounter: Int)
+                            programCounter: BigInt)
